+++ mne/io/brainvision/tests/test_brainvision.py
@@ -160,7 +160,7 @@ def test_ascii():
 def test_brainvision_data_highpass_filters():
     """Test reading raw Brain Vision files with amplifier filter settings."""
     # Homogeneous highpass in seconds (default measurement unit)
-    with pytest.warns(RuntimeWarning, match='parse triggers that'):
+    with pytest.warns(RuntimeWarning, match='will be dropped'):
         raw = _test_raw_reader(
             read_raw_brainvision, vhdr_fname=vhdr_highpass_path,
             montage=montage, eog=eog)
@@ -200,7 +200,7 @@ def test_brainvision_data_highpass_filters():
             read_raw_brainvision, vhdr_fname=vhdr_mixed_highpass_hz_path,
             montage=montage, eog=eog, event_id=event_id)
 
-    trigger_warning = ['parse triggers that' in str(ww.message)
+    trigger_warning = ['will be dropped' in str(ww.message)
                        for ww in w]
     lowpass_warning = ['different lowpass filters' in str(ww.message)
                        for ww in w]
@@ -278,7 +278,7 @@ def test_brainvision_data_partially_disabled_hw_filters():
             vhdr_fname=vhdr_partially_disabled_hw_filter_path,
             montage=montage, eog=eog)
 
-    trigger_warning = ['parse triggers that' in str(ww.message)
+    trigger_warning = ['will be dropped' in str(ww.message)
                        for ww in w]
     lowpass_warning = ['different lowpass filters' in str(ww.message)
                        for ww in w]
@@ -416,17 +416,17 @@ def test_events():
     events = raw._get_brainvision_events()
     events = events[events[:, 2] != event_id['Sync On']]
     assert_array_equal(events, [[486, 0, 253],
-                                [496, 1, 255],
-                                [1769, 1, 254],
-                                [1779, 1, 255],
-                                [3252, 1, 254],
-                                [3262, 1, 255],
-                                [4935, 1, 253],
-                                [4945, 1, 255],
-                                [5999, 1, 255],
-                                [6619, 1, 254],
-                                [6629, 1, 255],
-                                [7699, 1, 1]])
+                                [496, 0, 255],
+                                [1769, 0, 254],
+                                [1779, 0, 255],
+                                [3252, 0, 254],
+                                [3262, 0, 255],
+                                [4935, 0, 253],
+                                [4945, 0, 255],
+                                [5999, 0, 255],
+                                [6619, 0, 254],
+                                [6629, 0, 255],
+                                [7699, 0, 1]])
 
     # check that events are read and stim channel is synthesized correctly and
     # response triggers are shifted using the deprecated response_trig_shift.
@@ -435,18 +435,18 @@ def test_events():
                                    response_trig_shift=1000, event_id=event_id)
     events = raw._get_brainvision_events()
     assert_array_equal(events, [[486, 0, 253],
-                                [496, 1, 255],
-                                [1769, 1, 254],
-                                [1779, 1, 255],
-                                [3252, 1, 254],
-                                [3262, 1, 255],
-                                [4935, 1, 253],
-                                [4945, 1, 255],
-                                [5999, 1, 1255],
-                                [6619, 1, 254],
-                                [6629, 1, 255],
-                                [7629, 1, 5],
-                                [7699, 1, 1]])
+                                [496, 0, 255],
+                                [1769, 0, 254],
+                                [1779, 0, 255],
+                                [3252, 0, 254],
+                                [3262, 0, 255],
+                                [4935, 0, 253],
+                                [4945, 0, 255],
+                                [5999, 0, 1255],
+                                [6619, 0, 254],
+                                [6629, 0, 255],
+                                [7629, 0, 5],
+                                [7699, 0, 1]])
 
     # check that trig_shift_by_type works as well
     raw = read_raw_brainvision(vhdr_path, eog=eog,
@@ -455,21 +455,21 @@ def test_events():
                                event_id=event_id)
     events = raw._get_brainvision_events()
     assert_array_equal(events, [[486, 0, 253],
-                                [496, 1, 255],
-                                [1769, 1, 254],
-                                [1779, 1, 255],
-                                [3252, 1, 254],
-                                [3262, 1, 255],
-                                [4935, 1, 253],
-                                [4945, 1, 255],
-                                [5999, 1, 1255],
-                                [6619, 1, 254],
-                                [6629, 1, 255],
-                                [7629, 1, 5],
-                                [7699, 1, 2001]])
+                                [496, 0, 255],
+                                [1769, 0, 254],
+                                [1779, 0, 255],
+                                [3252, 0, 254],
+                                [3262, 0, 255],
+                                [4935, 0, 253],
+                                [4945, 0, 255],
+                                [5999, 0, 1255],
+                                [6619, 0, 254],
+                                [6629, 0, 255],
+                                [7629, 0, 5],
+                                [7699, 0, 2001]])
 
     # Check that we warn if a trigger is dropped
-    with pytest.warns(RuntimeWarning, match='to parse triggers'):
+    with pytest.warns(RuntimeWarning, match='will be dropped'):
         raw = read_raw_brainvision(vhdr_path)
     # check that events are read and stim channel is synthesized correctly and
     # response triggers are ignored.
@@ -478,16 +478,16 @@ def test_events():
     events = raw._get_brainvision_events()
     events = events[events[:, 2] != event_id['Sync On']]
     assert_array_equal(events, [[486, 0, 253],
-                                [496, 1, 255],
-                                [1769, 1, 254],
-                                [1779, 1, 255],
-                                [3252, 1, 254],
-                                [3262, 1, 255],
-                                [4935, 1, 253],
-                                [4945, 1, 255],
-                                [6619, 1, 254],
-                                [6629, 1, 255],
-                                [7699, 1, 1]])
+                                [496, 0, 255],
+                                [1769, 0, 254],
+                                [1779, 0, 255],
+                                [3252, 0, 254],
+                                [3262, 0, 255],
+                                [4935, 0, 253],
+                                [4945, 0, 255],
+                                [6619, 0, 254],
+                                [6629, 0, 255],
+                                [7699, 0, 1]])
 
     # Error handling of trig_shift_by_type
     pytest.raises(TypeError, read_raw_brainvision, vhdr_path, eog=eog,
@@ -526,32 +526,21 @@ def test_events():
                                trig_shift_by_type={'response': None},
                                event_id=event_id)
     events = raw._get_brainvision_events()
-    assert_array_equal(events, [[486, 0, 253],
-                                [496, 1, 255],
-                                [1769, 1, 254],
-                                [1779, 1, 255],
-                                [3252, 1, 254],
-                                [3262, 1, 255],
-                                [4935, 1, 253],
-                                [4945, 1, 255],
-                                [6619, 1, 254],
-                                [6629, 1, 255],
-                                [7629, 1, 5],
-                                [7699, 1, 1]])
 
-    # to handle the min duration = 1 of stim trig (re)construction ...
-    events = np.array([[486, 1, 253],
-                       [496, 1, 255],
-                       [1769, 1, 254],
-                       [1779, 1, 255],
-                       [3252, 1, 254],
-                       [3262, 1, 255],
-                       [4935, 1, 253],
-                       [4945, 1, 255],
-                       [6619, 1, 254],
-                       [6629, 1, 255],
-                       [7629, 1, 5],
-                       [7699, 1, 1]])
+    expected_events = np.array([[486, 0, 253],
+                                [496, 0, 255],
+                                [1769, 0, 254],
+                                [1779, 0, 255],
+                                [3252, 0, 254],
+                                [3262, 0, 255],
+                                [4935, 0, 253],
+                                [4945, 0, 255],
+                                [6619, 0, 254],
+                                [6629, 0, 255],
+                                [7629, 0, 5],
+                                [7699, 0, 1]])
+
+    assert_array_equal(events, expected_events)
 
     # Test that both trig_shift_by_type and event_id can be set
     read_raw_brainvision(vhdr_path, eog=eog, preload=False,
@@ -612,19 +601,23 @@ def test_brainvision_neuroone_export():
     assert len(raw.info['chs']) == 66
     assert raw.info['sfreq'] == 5000.
 
+
 @testing.requires_testing_data
 def test_read_vmrk_annotations():
     """Test annotations onsets are timestamps (+ validate some)"""
-    from mne.io.brainvision.brainvision import _read_vmrk_annotations
-    event_id = {'Sync On': 5, 'O  1': 1, 'R255': 255}
-    eog = ['HL', 'HR', 'Vb']
-    annotations = _read_vmrk_annotations(vmrk_path, event_id=event_id, sfreq=1000.0)
-    assert annotations.orig_time == 20131113161403794232
-    # validation_samples = [0, 1, 2, 3, 5, 7, 11, 13, 17, 19, 23, 29, 31, 37]
-    # expected = np.array([128.00, 217.00, 266.54, 602.00, 987.00, 1446.89,
-    #                      2199.99, 2585.50, 3343.98, 3730.01, 4511.53,
-    #                      5670.50, 6036.04, 7195.01])
-    # assert_array_almost_equal(annotations.onset[validation_samples],
-    #                           expected, decimal=2)
+    from mne.io.brainvision import read_annotations_brainvision
+    sfreq = 1000.0
+    annotations = read_annotations_brainvision(vmrk_path, sfreq=sfreq)
+    assert annotations.orig_time == 1384359243.794231
+    expected = np.array([486., 496., 1769., 1779., 3252., 3262., 4935., 4945.,
+                         5999., 6619., 6629., 7629., 7699.]) / sfreq
+    description = ['Stimulus/S253', 'Stimulus/S255', 'Stimulus/S254',
+                   'Stimulus/S255', 'Stimulus/S254', 'Stimulus/S255',
+                   'Stimulus/S253', 'Stimulus/S255', 'Response/R255',
+                   'Stimulus/S254', 'Stimulus/S255',
+                   'SyncStatus/Sync On', 'Optic/O  1']
+    assert_array_almost_equal(annotations.onset,
+                              expected, decimal=7)
+    assert_array_equal(annotations.description, description)
 
 run_tests_if_main()