+++ doc/whats_new.rst
@@ -71,6 +71,8 @@ Bug
 
 - Fix bug with reading events from BrainVision files by `Stefan Appelhoff`_
 
+- Don't use 2nd column of events in BrainVision to store duration but rather raw.annotations by `Alex Gramfort`_
+
 - Fix bug of not showing ERD's in baseline rescaled tfr topomaps if grads are combined by `Erkka Heinila`_
 
 - Fix bug with FIF I/O where strings were written in UTF-8 format instead of Latin-1 by `Eric Larson`_